pico: Rename csp.c and csp.h to pico_csp.c and pico_csp.h
Rename csp.c and csp.h to pico_csp.c and pico_csp.h, respectively, to
avoid confusion with libcsp's csp/csp.h and to clarify their
Pico-specific purpose.
