+++ pico/CMakeLists.txt
@@ -14,5 +14,5 @@ project(scsat1-rpi-pico)
 target_sources(app PRIVATE
 	src/main.c
 	src/temp.c
-	src/csp.c
+	src/pico_csp.c
 )
